Library/CloudStorage/Dropbox/Email Attachments Search/requirements.txt
Flask
python-dotenv
pdfminer.six
faiss-cpu
openai
requests
